running with synthetic

+++ content/modules/Datasets.py
@@ -392,7 +392,7 @@ def synthetic_dataset(path = 'data/', device='cpu'):
         graph_data[graph_idx]['node_list'] = torch.arange(0, graph_data[graph_idx]['num_nodes'])
         graph_data[graph_idx]['graph_idx'] = graph_idx # for reference
         # EDGE RELATED:
-        graph_data[graph_idx]['node_coordinates'] = graph_coords[torch.unique(data[torch.where(data[:, -1] == graph_idx)[0], 0]).to(torch.long)].to(torch.long)
+        graph_data[graph_idx]['node_coordinates'] = graph_coords[torch.unique(data[torch.where(data[:, -1] == graph_idx)[0], 0]).to(torch.long)].to(torch.float64)
         # Edge list - fully connected graphs thus perform possible combinations
         graph_data[graph_idx]['edge_list'] = torch.tensor(list(combinations(graph_data[graph_idx]['node_list'], 2))).to(torch.long)
         #graph_data[graph_idx]['edge_list'] = data[torch.where(data[:, -1] == graph_idx)[0]][:, :2]
